requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
python-dotenv
password_strength

pyyaml
uritemplate
inflection
drf-spectacular

# Database
psycopg[binary]==3.2.10

# Deployment
gunicorn==21.2.0


whitenoise